// website/src/sdk/tokens_pb.ts
export declare namespace Token {
	export type AsObject = {
		id: string,
		name: string,
		owner: string,
		ownerName: string,
		createdAt?: googleProtobufTimestamp.Timestamp.AsObject,
		expiresAt?: googleProtobufTimestamp.Timestamp.AsObject,
		lastUsedAt?: googleProtobufTimestamp.Timestamp.AsObject,
	}
}

export class Token extends jspb.Message {

	private static repeatedFields_ = [
		
	];

	constructor(data?: jspb.Message.MessageArray) {
		super();
		jspb.Message.initialize(this, data || [], 0, -1, Token.repeatedFields_, null);
	}


	getId(): string {return jspb.Message.getFieldWithDefault(this, 1, "");
	}

	setId(value: string): void {
		(jspb.Message as any).setProto3StringField(this, 1, value);
	}

	getName(): string {return jspb.Message.getFieldWithDefault(this, 2, "");
	}

	setName(value: string): void {
		(jspb.Message as any).setProto3StringField(this, 2, value);
	}

	getOwner(): string {return jspb.Message.getFieldWithDefault(this, 3, "");
	}

	setOwner(value: string): void {
		(jspb.Message as any).setProto3StringField(this, 3, value);
	}

	getOwnerName(): string {return jspb.Message.getFieldWithDefault(this, 4, "");
	}

	setOwnerName(value: string): void {
		(jspb.Message as any).setProto3StringField(this, 4, value);
	}

	getCreatedAt(): googleProtobufTimestamp.Timestamp {
		return jspb.Message.getWrapperField(this, googleProtobufTimestamp.Timestamp, 5);
	}

	setCreatedAt(value?: googleProtobufTimestamp.Timestamp): void {
		(jspb.Message as any).setWrapperField(this, 5, value);
	}

	getExpiresAt(): googleProtobufTimestamp.Timestamp {
		return jspb.Message.getWrapperField(this, googleProtobufTimestamp.Timestamp, 6);
	}

	setExpiresAt(value?: googleProtobufTimestamp.Timestamp): void {
		(jspb.Message as any).setWrapperField(this, 6, value);
	}

	getLastUsedAt(): googleProtobufTimestamp.Timestamp {
		return jspb.Message.getWrapperField(this, googleProtobufTimestamp.Timestamp, 7);
	}

	setLastUsedAt(value?: googleProtobufTimestamp.Timestamp): void {
		(jspb.Message as any).setWrapperField(this, 7, value);
	}

	serializeBinary(): Uint8Array {
		const writer = new jspb.BinaryWriter();
		Token.serializeBinaryToWriter(this, writer);
		return writer.getResultBuffer();
	}

	toObject(): Token.AsObject {
		let f: any;
		return {
			id: this.getId(),
			name: this.getName(),
			owner: this.getOwner(),
			ownerName: this.getOwnerName(),
			createdAt: (f = this.getCreatedAt()) && f.toObject(),
			expiresAt: (f = this.getExpiresAt()) && f.toObject(),
			lastUsedAt: (f = this.getLastUsedAt()) && f.toObject(),
		};
	}

	static serializeBinaryToWriter(message: Token, writer: jspb.BinaryWriter): void {
		const field1 = message.getId();
		if (field1.length > 0) {
			writer.writeString(1, field1);
		}
		const field2 = message.getName();
		if (field2.length > 0) {
			writer.writeString(2, field2);
		}
		const field3 = message.getOwner();
		if (field3.length > 0) {
			writer.writeString(3, field3);
		}
		const field4 = message.getOwnerName();
		if (field4.length > 0) {
			writer.writeString(4, field4);
		}
		const field5 = message.getCreatedAt();
		if (field5 != null) {
			writer.writeMessage(5, field5, googleProtobufTimestamp.Timestamp.serializeBinaryToWriter);
		}
		const field6 = message.getExpiresAt();
		if (field6 != null) {
			writer.writeMessage(6, field6, googleProtobufTimestamp.Timestamp.serializeBinaryToWriter);
		}
		const field7 = message.getLastUsedAt();
		if (field7 != null) {
			writer.writeMessage(7, field7, googleProtobufTimestamp.Timestamp.serializeBinaryToWriter);
		}
	}

	static deserializeBinary(bytes: Uint8Array): Token {
		var reader = new jspb.BinaryReader(bytes);
		var message = new Token();
		return Token.deserializeBinaryFromReader(message, reader);
	}

	static deserializeBinaryFromReader(message: Token, reader: jspb.BinaryReader): Token {
		while (reader.nextField()) {
			if (reader.isEndGroup()) {
				break;
			}
			const field = reader.getFieldNumber();
			switch (field) {
			case 1:
				const field1 = reader.readString()
				message.setId(field1);
				break;
			case 2:
				const field2 = reader.readString()
				message.setName(field2);
				break;
			case 3:
				const field3 = reader.readString()
				message.setOwner(field3);
				break;
			case 4:
				const field4 = reader.readString()
				message.setOwnerName(field4);
				break;
			case 5:
				const field5 = new googleProtobufTimestamp.Timestamp();
				reader.readMessage(field5, googleProtobufTimestamp.Timestamp.deserializeBinaryFromReader);
				message.setCreatedAt(field5);
				break;
			case 6:
				const field6 = new googleProtobufTimestamp.Timestamp();
				reader.readMessage(field6, googleProtobufTimestamp.Timestamp.deserializeBinaryFromReader);
				message.setExpiresAt(field6);
				break;
			case 7:
				const field7 = new googleProtobufTimestamp.Timestamp();
				reader.readMessage(field7, googleProtobufTimestamp.Timestamp.deserializeBinaryFromReader);
				message.setLastUsedAt(field7);
				break;
			default:
				reader.skipField();
				break;
			}
		}
		return message;
	}

}